[panels] Fix Bugzilla Backlog with correct fields
This panel was using an old date as time series field. This has
been updated to use `grimoire_creation_date`.

`project_name` field is also replaced by `project` field, which is
the common field that must be used among indices for compatibility
with the rest of the tools in Grimoire paltform.

Old fields have been removed from index pattern.