[tests] Remove from test checking the ocean-unique-id field
This field is not used anymore in GrimoireELK and it is not
used in the panels. It just exists in the index patterns because
it existed in GrimoireELK enriched indexes.

This field has been removed with the arthur support. Now the unique
id for all items is always uuid, the unique id coming from arthur.