[overview] Include for studies vis a "_studies_" in the id and title
git pair programming study is normaly not present, so it must not be
included in Overview if the default case. This change modifies the
id and title for visualizations of this study to specify that it is
a study vis:

git_pair_programming_title -> git_study_pair_programming_title

This solution can be used in general so kidash knows what vis are
for studies, and to decide if they must be included.