[docs] Add Jekyll based documentation
Add files and configuration needed to build the project
using GitHub Pages. From now on, we will use collections
of panels to group set of panels that are somewhat related.
In this commit we include a `sample_collection` to be used
in the future as a simple template for new collections and
also all the files needed to document current Sigils panels
as a collection.