[affiliations] Set limit author entries due to performance issues
The current limit in affiliations page was set to 500000.
This limit is too high and can cause problems since Kibana
loads all these data in memory.

The new limit has been set to 1000. Most browsers will be
able to handle this amount of information at the same time.

This commit also includes a visualization markdown with help
about Affiliations panel and why the limit has been set
to this value.