[cocom] Update cocom panels
This code updates the cocom panels to prevent visualizing
credentials included in the origin of a git repository. The
visualizations with titles "Repositories" and the "All Files:
Code complexity, LOC and ratios" have been updated and now
use the field `repo_url` (which doesn't expose credentials)
instead of `origin`.

Note that the origin and the credentials can be visualized by
inspecting the raw or/and enriched index.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>