[github-closed-events] Include submitters and reporters in dashboard
This code improves the visualizations for the dashboard around the
connections between the issues closed by pull requests. Now, the
information about the issue reporters and pull request submitters
is shown on the dashboard. In particular, it is possible to know the
global numbers of submitters and reporters, the details for each issue,
and see how many issues have been fixed by a given submitters
and the organization he belongs to.

Details about single contributors fixing issues have been removed.

Documentation has been updated.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>