[gerrit] Update index pattern
This index pattern has been updated beacuse there were some missing
fields, such as `demography_*` fields.
This caused Kibana to hide the fields even when the data is in the
index.

In addition to the these fields, there are some new ones too
after the addition of `demographics_contribution` study in
grimoire-elk repository.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>