[gerrit] Update index patterns with last Verified approval value
New field: `changeset_status_value_verified`. It is defined as the
value of the last "Verified" approval made by someone else than the
author for a given changeset, it can be `-2`, `-1`, `1` or `2`.

Both `gerrit` and `demographics_contribution` index patterns have
been fixed as last version contained unnecessary fields which were
not part of the index pattern before last time it was changed.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>